Updated ShoppingCart Controller

+++ src/main/java/se/iths/projektarbetekomplexjava/controller/ShoppingCartController.java
@@ -6,6 +6,7 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import se.iths.projektarbetekomplexjava.entity.*;
+import se.iths.projektarbetekomplexjava.exception.NotAuthorizedException;
 import se.iths.projektarbetekomplexjava.exception.NotFoundException;
 import se.iths.projektarbetekomplexjava.repository.*;
 import se.iths.projektarbetekomplexjava.service.BookService;
@@ -44,6 +45,9 @@ public class ShoppingCartController {
     @PutMapping("/addbooks/bookid/{bookid}/username/{username}/{qty}")
     public ResponseEntity<CartItem> addBooksToCart(@PathVariable int qty, @PathVariable String username, @PathVariable Long bookid) {
         Customer customer = customerRepository.findByUsername(username);
+        if (!customer.getLoggedInCustomer().isLoggedIn()){
+            throw new NotAuthorizedException("You need to login in order to add a book to the cart.");
+        }
         ShoppingCart shoppingCart = customer.getShoppingCart();
         Book book;
         book = bookService.findByBookId(bookid).orElseThrow(EntityNotFoundException::new);
@@ -60,19 +64,25 @@ public class ShoppingCartController {
     @PutMapping("/updateCartItem/{cartid}/quantity/{qty}")
     public ResponseEntity<CartItem> updateCartItem(@PathVariable Long cartid, @PathVariable int qty) {
         CartItem cartItem = cartItemService.findById(cartid);
+        Customer customer = cartItem.getShoppingCart().getCustomer();
+        if (!customer.getLoggedInCustomer().isLoggedIn()){
+            throw new NotAuthorizedException("You need to login in order to add a book to the cart.");
+        }
         cartItem.setQty(cartItem.getQty() + qty);
         cartItemService.updateCartItem(cartItem);
         shoppingCartService.updateShoppingCart(cartItem.getShoppingCart());
         return new ResponseEntity<>(cartItem, HttpStatus.OK);
-
     }
 
     @PutMapping("/removeBookFromCart/{cartid}/quantity/{qty}")
     public ResponseEntity<CartItem> removeBookFromCartItem(@PathVariable Long cartid, @PathVariable int qty) {
         CartItem cartItem = cartItemService.findById(cartid);
+        Customer customer = cartItem.getShoppingCart().getCustomer();
+        if (!customer.getLoggedInCustomer().isLoggedIn()){
+            throw new NotAuthorizedException("You need to login in order to add a book to the cart.");
+        }
         if(cartItem.getQty() < qty) {
             throw new NotFoundException("you do not have : " + qty + " books in your cart");
-
         }
         else if(cartItem.getQty() == qty){
             try {
@@ -81,7 +91,6 @@ public class ShoppingCartController {
             finally {
                 cartItemService.removeCartItem(cartItemService.findById(cartid));
             }
-
         }
 
         cartItem.setQty(cartItem.getQty() - qty);
@@ -91,19 +100,24 @@ public class ShoppingCartController {
         return new ResponseEntity<>(cartItem, HttpStatus.OK);
     }
 
-    @DeleteMapping("/removebookfromcart/{id}")
+    @DeleteMapping("/removeCart/{id}")
     public ResponseEntity<Void> removeBookFromCart(@PathVariable Long id) {
+        CartItem cartItem = cartItemService.findById(id);
+        Customer customer = cartItem.getShoppingCart().getCustomer();
+        if (!customer.getLoggedInCustomer().isLoggedIn()){
+            throw new NotAuthorizedException("You need to login in order to add a book to the cart.");
+        }
         cartItemService.removeCartItem(cartItemService.findById(id));
         return new ResponseEntity<>(HttpStatus.RESET_CONTENT);
     }
 
-    @PutMapping("/updateShoppingCart/username/{uname}")
-    public ResponseEntity<ShoppingCart> updateShoppingCart(@PathVariable String uname) {
-        Customer customer = customerRepository.findByUsername(uname);
-        ShoppingCart shoppingCart = customer.getShoppingCart();
-        //  List<CartItem> cartItemList=cartItemService.findByShoppingCart(shoppingCart);
-        shoppingCartService.updateShoppingCart(shoppingCart);
-        return new ResponseEntity<>(shoppingCart, HttpStatus.ACCEPTED);
-
-    }
+//    @PutMapping("/updateShoppingCart/username/{uname}")
+//    public ResponseEntity<ShoppingCart> updateShoppingCart(@PathVariable String uname) {
+//        Customer customer = customerRepository.findByUsername(uname);
+//        ShoppingCart shoppingCart = customer.getShoppingCart();
+//        //  List<CartItem> cartItemList=cartItemService.findByShoppingCart(shoppingCart);
+//        shoppingCartService.updateShoppingCart(shoppingCart);
+//        return new ResponseEntity<>(shoppingCart, HttpStatus.ACCEPTED);
+//
+//    }
 }
